country grid max 30, overview graph scale, cum choose all, more colours
The country grid on mobile is now capped at 30. The mobile issues I were experiencing were specific to one build so not sur what was going on there.
Added Scale to the country overview graph.
On the cumulative graph you can now select all. made the code in this component DRYer